backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.4
djangorestframework==3.16.0
faiss-cpu==1.11.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.4.4
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
ollama==0.5.1
openai==1.93.0
orjson==3.10.18
packaging==24.2
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdf==5.7.0
PyPDF2==3.0.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
zstandard==0.23.0
gunicorn
